prevent route

+++ routes/web.php
@@ -21,13 +21,16 @@ Route::get('logout', [AuthController::class, 'doLogout'])->name('auth.doLogout')
 Route::get('/user/create', [UserController::class, 'create'])->name('user.create');
 Route::post('/user', [UserController::class, 'store'])->name('user.store');
 
+// route to forbidden
+Route::get('/users/forbidden', [UserController::class, 'goForbidden'])->name('user.forbidden');
+
 // routes for posts crud
-Route::get('/posts', [PostController::class, 'index'])->name('posts.home');
-Route::get('/posts/create', [PostController::class, 'create'])->name('posts.create');
-Route::post('/posts', [PostController::class, 'store'])->name('posts.store');
-Route::get('/posts/show/{id}', [PostController::class, 'show'])->name('posts.show');
-Route::get('/posts/edit/{id}', [PostController::class, 'edit'])->name('posts.edit');
-Route::put('/posts/update/{id}', [PostController::class, 'update'])->name('posts.update');
-Route::delete('/posts/delete/{id}', [PostController::class, 'destroy'])->name('posts.delete');
+Route::get('/posts', [PostController::class, 'index'])->name('posts.home')->middleware('auth');
+Route::get('/posts/create', [PostController::class, 'create'])->name('posts.create')->middleware('auth');
+Route::post('/posts', [PostController::class, 'store'])->name('posts.store')->middleware('auth');
+Route::get('/posts/show/{id}', [PostController::class, 'show'])->name('posts.show')->middleware('auth');
+Route::get('/posts/edit/{id}', [PostController::class, 'edit'])->name('posts.edit')->middleware('auth');
+Route::put('/posts/update/{id}', [PostController::class, 'update'])->name('posts.update')->middleware('auth');
+Route::delete('/posts/delete/{id}', [PostController::class, 'destroy'])->name('posts.delete')->middleware('auth');
